+++ apps/www/CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+-   No longer setting `document.domain`.
+
 ### Security
 
 ## [2.0.6] - 2023-02-29